pyproject.toml
[tool.poetry]
name = "diafilm2"
version = "0.1.0"
description = "«Δиа Фильм²» – это проект, который разрабатывался для курса Яндекс Практикум"
authors = ["mbrav <mbrav@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0"
Pillow = "^9.0.0"
djangorestframework = "^3.13.1"
django-filter = "^21.1"
django-environ = "^0.8.1"
django-rest-swagger = "^2.2.0"
sorl-thumbnail = "^12.7.0"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.2.4"
pytest-django = "^4.5.2"
Markdown = "^3.3.6"
flake8 = "^4.0.1"
autopep8 = "^1.6.0"
transliterate = "^1.10.2"
beautifulsoup4 = "^4.10.0"
tqdm = "^4.62.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
